+++ edsteva/probes/condition/completeness_predictors/per_visit.py
@@ -40,6 +40,7 @@ def compute_completeness_predictor_per_visit(
     source_systems: Union[bool, List[str]],
     length_of_stays: List[float],
     age_ranges: List[int],
+    gender_source_values: Union[bool, str, Dict[str, str]],
     provenance_sources: Union[bool, str, Dict[str, str]],
     stay_sources: Union[bool, str, Dict[str, str]],
     drg_sources: Union[bool, str, Dict[str, str]],
@@ -75,7 +76,11 @@ def compute_completeness_predictor_per_visit(
         data=data,
     )
     self.care_site_relationship = care_site_relationship
-    person = prepare_person(data) if age_ranges else None
+    person = (
+        prepare_person(data, gender_source_values)
+        if (age_ranges or gender_source_values)
+        else None
+    )
     cost = prepare_cost(data, drg_sources) if drg_sources else None
 
     visit_occurrence = prepare_visit_occurrence(
@@ -259,6 +264,7 @@ def get_uf_visit(
                         "age_range",
                         "drg_source",
                         "condition_type",
+                        "gender_source_value",
                     ]
                 )
             )