+++ edsteva/probes/note/completeness_predictors/per_visit.py
@@ -42,6 +42,7 @@ def compute_completeness_predictor_per_visit(
     length_of_stays: List[float],
     note_types: Union[str, Dict[str, str]],
     age_ranges: List[int],
+    gender_source_values: Union[bool, str, Dict[str, str]],
     condition_types: Union[bool, str, Dict[str, str]],
     provenance_sources: Union[bool, str, Dict[str, str]],
     stay_sources: Union[bool, str, Dict[str, str]],
@@ -74,7 +75,11 @@ def compute_completeness_predictor_per_visit(
         data=data,
     )
     self.care_site_relationship = care_site_relationship
-    person = prepare_person(data) if age_ranges else None
+    person = (
+        prepare_person(data, gender_source_values)
+        if (age_ranges or gender_source_values)
+        else None
+    )
     cost = prepare_cost(data, drg_sources) if drg_sources else None
 
     visit_occurrence = prepare_visit_occurrence(